requirements.txt
openai>=1.0.0
Pillow>=9.0.0
numpy>=1.22.0
